+++ examples/03_native.zig
@@ -44,6 +44,7 @@ fn onWindowEvent(ctx: ?*anyopaque, event: zgpui.platform.WindowEvent) void {
             state.should_exit = true;
         },
         .a11y_press => {},
+        .a11y_adjust => {},
     }
 }